small refactoring and improve of kdoc
